services/notification_service/requirements.txt
django==5.1.0 
djangorestframework==3.15.2
psycopg2-binary==2.9.9 
gunicorn==23.0.0 
django-environ==0.11.2
pika==1.3.2
python-consul==1.1.0
django-cors-headers==4.4.0
